+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: `BackingCard`: Add gradient on truncated description.
 - Fix: `BackingCard`: New styles for `disabled` state.
+- Fix: `BackingCard`: Better margin when no image.
 
 ## [8.0.0] - 2022-02-07